+++ NEWS.md
@@ -10,9 +10,13 @@
   that all elements of the vector are named. (#70)
 
 ## Breaking Changes
+- The default value of the `optional` argument in `assert_date_vector()`,
+`assert_list_of()`, and `assert_s3_class()` was changed from `TRUE` to `FALSE`
+to make the default behavior consistent. (#87)
 
 ## Documentation
   - New section in programming strategy regarding comments (#71)
+
 
 # admiraldev 0.2.0